Create WowserBrowserV1
Create an app that is a basic Google Chrome web browser clone with one amazing little feature, an all in one A.i. Floatbox that can be hidden or brought out and put to use, its opacity at 75%. This awesome little box packs an explosion of user-based tools and re-generation features in which the a.i. Scans a webpage in the form of a screenshot then allows user to edit/create/reimagine certain aspects of that screenshot/webpage for example google images would work as a normal with the webpage able to search actual web based images from the google search bar  pulling images from https://google.com) but when the small black triangle floating button in the top left corner of your